WIP: move validate code from public API into common Impl module
